When telling clients to add a map to the rotation, the position should be sent as a short, not as a byte, in case map lists are larger than 255 maps.

+++ src/cl_main.cpp
@@ -2303,7 +2303,7 @@ void CLIENT_ProcessCommand( LONG lCommand, BYTESTREAM_s *pByteStream )
 						case UPDATE_MAPROTATION_ADDMAP:
 						{
 							const char *pszMapName = pByteStream->ReadString();
-							int position = pByteStream->ReadByte();
+							int position = pByteStream->ReadShort();
 							ULONG ulMinPlayers = pByteStream->ReadByte();
 							ULONG ulMaxPlayers = pByteStream->ReadByte();
 							MAPROTATION_AddMap( pszMapName, position, ulMinPlayers, ulMaxPlayers, true );
